Final/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
redis==5.0.1
elasticsearch==8.11.0
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
python-multipart==0.0.6
python-dotenv==1.0.0

pillow==10.1.0
pytesseract==0.3.10
pypdf2==3.0.1
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0